Add several features
Autogenerate an UID to link the timer codeblock with the state
`_timedUID` in the code block if deleted a new id will be generated

Auto Log on Stop
Settings to rename buttons
Total Time Calculation
Start and End Datetimes

Timer state is kept in the backgroud (while obsidian is open)
só you can start a timer in a note, close it and when reopened
the timer will continue running with the correct duration.

Fixed multiple timers running at the same time. (same or diferent notes)